+++ inc/views/class-statsbookreviews.php
@@ -246,9 +246,7 @@ class StatsBookReviews {
 			}
 
 			// set the reviewer's average
-//			if ( isset( $this->responseByUid[ $val['info1'] ] ) && isset( $this->responseByUid[ $val['info1'] ][ $val['id'] ])) {
 			$this->responseByUid[ $val['info1'] ][ $val['id'] ]['avg_score'] = round( $sum / $count, 2 );
-//			}
 		}
 
 		// set the average score and total reviews for each book
